Fixed up network stack and drivers for x86_64.
* Various compilation fixes.
* Fixes to the FreeBSD compatibility layer (from comparing the x86-
  specific bits with the equivalent amd64 sources in FreeBSD).
* Compile all the Ethernet drivers except for sis900 and wb840, these
  require a bit more work to fix (will file a ticket soon). Tested
  ipro1000 and rtl81xx, no issues.

+++ headers/private/net/net_buffer.h
@@ -85,9 +85,9 @@ struct net_buffer_module_info {
 
 	status_t		(*direct_access)(net_buffer* buffer, uint32 offset,
 						size_t bytes, void** _data);
-	status_t 		(*read)(net_buffer* buffer, uint32 offset, void* data,
+	status_t 		(*read)(net_buffer* buffer, size_t offset, void* data,
 						size_t bytes);
-	status_t		(*write)(net_buffer* buffer, uint32 offset,
+	status_t		(*write)(net_buffer* buffer, size_t offset,
 						const void* data, size_t bytes);
 
 	int32			(*checksum)(net_buffer* buffer, uint32 offset, size_t bytes,
